leave paginated fetch out of rtk query for now
